requirements.txt
django==4.0.1
djangorestframework==3.14.0
django-extensions==3.2.3